+++ src/main/java/com/liquor_store/order/service/OrderService.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 public interface OrderService {
     List<OrderDto> getAllOrders();
-    OrderDto getOrderById(Long id);
+    OrderDto getOrderById(Integer id);
     OrderDto createOrder(OrderDto orderDto);
-    OrderDto updateOrder(Long id, OrderDto orderDto);
-    void deleteOrder(Long id);
+    OrderDto updateOrder(Integer id, OrderDto orderDto);
+    void deleteOrder(Integer id);
 }
