+++ src/main/java/com/liquor_store/transaction/repository/SalesTransactionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SalesTransactionRepository extends JpaRepository<SalesTransaction, Long> {
+public interface SalesTransactionRepository extends JpaRepository<SalesTransaction, Integer> {
 }
